+++ src/LanguageUtils.js
@@ -8,10 +8,6 @@ class LanguageUtil {
   constructor(options) {
     this.options = options;
 
-    // temporal backwards compatibility WHITELIST REMOVAL
-    this.whitelist = this.options.supportedLngs || false;
-    // end temporal backwards compatibility WHITELIST REMOVAL
-
     this.supportedLngs = this.options.supportedLngs || false;
     this.logger = baseLogger.create('languageUtils');
   }
@@ -62,17 +58,6 @@ class LanguageUtil {
 
     return this.options.cleanCode || this.options.lowerCaseLng ? code.toLowerCase() : code;
   }
-
-  // temporal backwards compatibility WHITELIST REMOVAL
-  isWhitelisted(code) {
-    this.logger.deprecate(
-      'languageUtils.isWhitelisted',
-      'function "isWhitelisted" will be renamed to "isSupportedCode" in the next major - please make sure to rename it\'s usage asap.',
-    );
-
-    return this.isSupportedCode(code);
-  }
-  // end temporal backwards compatibility WHITELIST REMOVAL
 
   isSupportedCode(code) {
     if (this.options.load === 'languageOnly' || this.options.nonExplicitSupportedLngs) {